+++ rustfs/src/table_catalog/tests.rs
@@ -3,7 +3,9 @@ use super::identifier::{
     default_table_lifecycle_path, default_table_marker_path, default_table_root_prefix, is_valid_table_metadata_file_name,
     namespace_name_from_marker_path, table_name_from_marker_path, validate_object_mutation,
 };
-use super::test_support::{BlockingObjectPublication, TestCatalogObjectBackend, UnserializedTestPublication};
+use super::test_support::{
+    BlockingObjectPublication, NoopTableCatalogStore, TestCatalogObjectBackend, UnserializedTestPublication,
+};
 use super::*;
 use datafusion::{
     arrow::{
@@ -220,200 +222,6 @@ fn catalog_object_listing_rejects_missing_or_stalled_continuation_tokens() {
         catalog_list_next_continuation(&mut seen, true, Some("next".to_string())),
         Err(TableCatalogStoreError::Internal(message)) if message.contains("did not advance")
     );
-}
-
-struct NoopTableCatalogStore;
-
-#[async_trait::async_trait]
-impl TableCatalogStore for NoopTableCatalogStore {
-    async fn get_table_bucket(&self, _table_bucket: &str) -> TableCatalogStoreResult<Option<TableBucketEntry>> {
-        Ok(None)
-    }
-
-    async fn put_table_bucket(&self, _entry: TableBucketEntry) -> TableCatalogStoreResult<()> {
-        Ok(())
-    }
-
-    async fn create_namespace(&self, _entry: NamespaceEntry) -> TableCatalogStoreResult<()> {
-        Ok(())
-    }
-
-    async fn list_namespaces(&self, _table_bucket: &str) -> TableCatalogStoreResult<Vec<NamespaceEntry>> {
-        Ok(Vec::new())
-    }
-
-    async fn get_namespace(&self, _table_bucket: &str, _namespace: &str) -> TableCatalogStoreResult<Option<NamespaceEntry>> {
-        Ok(None)
-    }
-
-    async fn drop_namespace(&self, _table_bucket: &str, _namespace: &str) -> TableCatalogStoreResult<()> {
-        Ok(())
-    }
-
-    async fn create_table(&self, _entry: TableEntry) -> TableCatalogStoreResult<()> {
-        Ok(())
-    }
-
-    async fn register_table(&self, _entry: TableEntry) -> TableCatalogStoreResult<()> {
-        Ok(())
-    }
-
-    async fn register_table_with_publication(
-        &self,
-        entry: TableEntry,
-        publication: &(dyn TableCommitPublication + Sync),
-    ) -> TableCatalogStoreResult<()> {
-        publication.begin_table_bucket(&entry.table_bucket).await?;
-        if !publication.holds_table_bucket(&entry.table_bucket) {
-            return Err(TableCatalogStoreError::Internal(
-                "table registration requires a table-bucket publication fence".to_string(),
-            ));
-        }
-        let _publication_completion = TableCommitPublicationCompletion::new(publication);
-        publication
-            .prepare(&entry.table_bucket, &entry.namespace, &entry.table)
-            .await?;
-        if !publication.holds_table(&entry.table_bucket, &entry.namespace, &entry.table) {
-            return Err(TableCatalogStoreError::Internal(
-                "table registration requires a table publication fence".to_string(),
-            ));
-        }
-        self.register_table(entry).await
-    }
-
-    async fn list_tables(&self, _table_bucket: &str, _namespace: &str) -> TableCatalogStoreResult<Vec<TableEntry>> {
-        Ok(Vec::new())
-    }
-
-    async fn list_all_tables(&self, _table_bucket: &str) -> TableCatalogStoreResult<Vec<TableEntry>> {
-        Ok(Vec::new())
-    }
-
-    async fn load_table(
-        &self,
-        _table_bucket: &str,
-        _namespace: &str,
-        _table: &str,
-    ) -> TableCatalogStoreResult<Option<TableEntry>> {
-        Ok(None)
-    }
-
-    async fn commit_table(&self, request: TableCommitRequest) -> TableCatalogStoreResult<TableCommitResult> {
-        let table = TableEntry {
-            version: TABLE_CATALOG_ENTRY_VERSION,
-            table_bucket: request.table_bucket,
-            namespace: request.namespace,
-            table: request.table,
-            table_id: "table-id".to_string(),
-            table_uuid: "table-uuid".to_string(),
-            format: "ICEBERG".to_string(),
-            format_version: 2,
-            warehouse_location: "s3://analytics/tables/table-id".to_string(),
-            metadata_location: request.new_metadata_location.clone(),
-            version_token: "token-v2".to_string(),
-            generation: 2,
-            state: TableCatalogEntryState::Active,
-            properties: BTreeMap::new(),
-            created_at: None,
-            updated_at: None,
-        };
-        let commit_log = CommitLogEntry {
-            version: TABLE_CATALOG_ENTRY_VERSION,
-            commit_id: request.commit_id,
-            idempotency_key: request.idempotency_key,
-            table_id: table.table_id.clone(),
-            operation: request.operation,
-            expected_version_token: request.expected_version_token,
-            new_version_token: table.version_token.clone(),
-            previous_metadata_location: request.expected_metadata_location,
-            new_metadata_location: table.metadata_location.clone(),
-            requirements: request.requirements,
-            status: CommitLogStatus::Committed,
-            writer: request.writer,
-            created_at: None,
-            updated_at: None,
-        };
-
-        Ok(TableCommitResult { table, commit_log })
-    }
-
-    async fn commit_table_with_publication(
-        &self,
-        request: TableCommitRequest,
-        publication: &(dyn TableCommitPublication + Sync),
-    ) -> TableCatalogStoreResult<TableCommitResult> {
-        publication
-            .prepare(&request.table_bucket, &request.namespace, &request.table)
-            .await?;
-        if !publication.holds_table(&request.table_bucket, &request.namespace, &request.table) {
-            return Err(TableCatalogStoreError::Internal(
-                "table commit requires a table publication fence".to_string(),
-            ));
-        }
-        let _publication_completion = TableCommitPublicationCompletion::new(publication);
-        self.commit_table(request).await
-    }
-
-    async fn drop_table(&self, _table_bucket: &str, _namespace: &str, _table: &str) -> TableCatalogStoreResult<()> {
-        Ok(())
-    }
-
-    async fn create_view(&self, _entry: ViewEntry) -> TableCatalogStoreResult<()> {
-        Ok(())
-    }
-
-    async fn list_views(&self, _table_bucket: &str, _namespace: &str) -> TableCatalogStoreResult<Vec<ViewEntry>> {
-        Ok(Vec::new())
-    }
-
-    async fn load_view(&self, _table_bucket: &str, _namespace: &str, _view: &str) -> TableCatalogStoreResult<Option<ViewEntry>> {
-        Ok(None)
-    }
-
-    async fn replace_view(&self, request: ViewCommitRequest) -> TableCatalogStoreResult<ViewCommitResult> {
-        Ok(ViewCommitResult {
-            view: ViewEntry {
-                version: TABLE_CATALOG_ENTRY_VERSION,
-                table_bucket: request.table_bucket,
-                namespace: request.namespace,
-                view: request.view,
-                view_id: "view-id".to_string(),
-                view_uuid: "view-uuid".to_string(),
-                format: "ICEBERG_VIEW".to_string(),
-                format_version: 1,
-                warehouse_location: "s3://analytics/views/view-id".to_string(),
-                metadata_location: request.new_metadata_location,
-                version_token: "token-v2".to_string(),
-                generation: 2,
-                state: TableCatalogEntryState::Active,
-                properties: BTreeMap::new(),
-                created_at: None,
-                updated_at: None,
-            },
-        })
-    }
-
-    async fn drop_view(&self, _table_bucket: &str, _namespace: &str, _view: &str) -> TableCatalogStoreResult<()> {
-        Ok(())
-    }
-
-    async fn get_commit_by_id(
-        &self,
-        _table_bucket: &str,
-        _table_id: &str,
-        _commit_id: &str,
-    ) -> TableCatalogStoreResult<Option<CommitLogEntry>> {
-        Ok(None)
-    }
-
-    async fn get_commit_by_idempotency_key(
-        &self,
-        _table_bucket: &str,
-        _table_id: &str,
-        _idempotency_key: &str,
-    ) -> TableCatalogStoreResult<Option<CommitLogEntry>> {
-        Ok(None)
-    }
 }
 
 #[tokio::test]
